rename bareos-fd to bareos-client, edit comments
Rename to clarify usage.
Deployment of cfg files checks for homebrew's etc path.
