+++ Formula/bareos-client.rb
@@ -1,0 +1,74 @@
+class BareosClient < Formula
+  desc "Client for Bareos (Backup Archiving REcovery Open Sourced)"
+  homepage "https://www.bareos.org/"
+  url "https://github.com/bareos/bareos/archive/Release/17.2.4.tar.gz"
+  sha256 "4c443539012cf5ebb0fdb18878e604e82b951e6429c618acd18762f3c5724799"
+
+  depends_on "openssl"
+  depends_on "readline"
+
+  def install
+    system "./configure", "--prefix=#{prefix}",
+                          "--with-working-dir=#{var}/lib/bareos",
+                          "--with-archivedir=#{var}/bareos",
+                          "--with-confdir=#{etc}/bareos",
+                          "--with-configtemplatedir=#{lib}/bareos/defaultconfigs",
+                          "--with-scriptdir=#{lib}/bareos/scripts",
+                          "--with-plugindir=#{lib}/bareos/plugins",
+                          "--with-fd-password=XXX_REPLACE_WITH_CLIENT_PASSWORD_XXX",
+                          "--with-mon-fd-password=XXX_REPLACE_WITH_CLIENT_MONITOR_PASSWORD_XXX",
+                          "--with-basename=XXX_REPLACE_WITH_LOCAL_HOSTNAME_XXX",
+                          "--with-hostname=XXX_REPLACE_WITH_LOCAL_HOSTNAME_XXX",
+                          "--with-python",
+                          "--enable-client-only"
+
+    system "make"
+    # The file platforms/osx/Makefile is intended for other environment (not homebrew)
+    # and would break the build process.
+    # Therefore it is removed until this has been fixed upstream,
+    # see https://bugs.bareos.org/view.php?id=900
+    rm "platforms/osx/Makefile"
+    system "make", "install"
+  end
+
+  def post_install
+    # If no configuration files are present,
+    # deploy them (copy them and replace variables).
+    unless File.exist?("#{etc}/bareos/bareos-fd.d")
+      system("#{lib}/bareos/scripts/bareos-config deploy_config #{lib}/bareos/defaultconfigs #{etc}/bareos bareos-fd")
+      system("#{lib}/bareos/scripts/bareos-config deploy_config #{lib}/bareos/defaultconfigs #{etc}/bareos bconsole")
+    end
+  end
+
+  plist_options :startup => true
+
+  def plist; <<~EOS
+    <?xml version="1.0" encoding="UTF-8"?>
+    <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+    <plist version="1.0">
+      <dict>
+        <key>Label</key>
+        <string>#{plist_name}</string>
+        <key>ProgramArguments</key>
+        <array>
+          <string>#{opt_bin}/bareos-fd</string>
+          <string>-f</string>
+        </array>
+        <key>StandardOutPath</key>
+        <string>#{var}/run/bareos-fd.log</string>
+        <key>StandardErrorPath</key>
+        <string>#{var}/run/bareos.log</string>
+        <key>RunAtLoad</key>
+        <true/>
+      </dict>
+    </plist>
+  EOS
+  end
+
+  test do
+    # Check if bareos-fd starts at all.
+    assert_match version.to_s, shell_output("#{bin}/bareos-fd -? 2>&1", 1)
+    # Check if the configuration is valid.
+    system "#{bin}/bareos-fd", "-t"
+  end
+end
